backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
requests==2.31.0
openai>=1.12.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
cryptography==42.0.0
tenacity==8.2.3
structlog==24.1.0
Pillow==10.2.0
aiofiles==23.2.1
apscheduler==3.10.4
